UpdatesActivity: allow sorting from oldest to newest
We require the user to install builds in order.
Assume we have 3 updates: A, B and C.
C depends on B, B depends on A.

Old behavior:
 - sort from newest to oldest
 - show single update
 - result: show update C, skip A and B

New behavior:
 - sort from oldest to newest
 - show single update
 - result: show update A, skip B and C

The old behavior always offered the latest build. If these are incremental
OTAs, it would not work as the system expects a specific state.

With the new behavior, we do not skip to the latest build, but go step by step
from oldest to newest.

This will ensure that builds are installed in order and satisfy dependencies.

Change-Id: Ic3f446280f98a2dd15182e713cced9781aa4371f
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>